test: Updated code to better accomodate latest fixes in the templates page,  removed redundant old alert module code files, updated playwright yml file
