+++ .github/workflows/playwright.yml
@@ -141,7 +141,7 @@ jobs:
             "unflattened.spec.js",
             "usersOrg.spec.js",
             "websocket.spec.js",
-            "alertsNew.spec.js",
+            "alerts.spec.js",
              # "remotepipeline.spec.js",
 
           ]
